requirements.txt
python-telegram-bot==21.10
SQLAlchemy==2.0.46
psycopg[binary]==3.2.13
alembic==1.14.1
python-dotenv==1.0.1
APScheduler==3.10.4
pydantic==2.12.5
pydantic-settings==2.13.1
fastapi==0.115.11
uvicorn[standard]==0.34.0
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
jinja2==3.1.6
python-multipart==0.0.20
